feat: 역 이전에 구간 먼저 삭제
역을 삭제하기 위해서는 해당 역을 사용중인 구간을 먼저 삭제해야만 삭제 가능하도록 구현하였습니다.

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -2,9 +2,13 @@ package wooteco.subway.service;
 
 import static java.util.stream.Collectors.toUnmodifiableList;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import org.springframework.stereotype.Service;
+import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
+import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
@@ -16,9 +20,11 @@ public class StationService {
     private static final int DELETE_FAIL = 0;
 
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse save(StationRequest request) {
@@ -37,10 +43,31 @@ public class StationService {
     }
 
     public void deleteById(Long id) {
-        final int isDeleted = stationDao.deleteById(id);
+        Set<Long> stationIds = new HashSet<>();
+        final List<Section> sections = sectionDao.findAll();
 
+        getStationIds(stationIds, sections);
+        validateDeletable(id, stationIds);
+
+        final int isDeleted = stationDao.deleteById(id);
         if (isDeleted == DELETE_FAIL) {
             throw new NotExistException("존재하지 않는 지하철 역입니다.");
+        }
+    }
+
+    private void getStationIds(Set<Long> stationIds, List<Section> sections) {
+        for (Section section : sections) {
+            final Long upStationId = section.getUpStationId();
+            final Long downStationId = section.getDownStationId();
+
+            stationIds.add(upStationId);
+            stationIds.add(downStationId);
+        }
+    }
+
+    private void validateDeletable(Long id, Set<Long> stationIds) {
+        if (stationIds.contains(id)) {
+            throw new IllegalArgumentException("구간을 먼저 삭제해주세요.");
         }
     }
 }
